requirements.txt
yt-dlp
youtube-transcript-api
requests
python-dotenv
streamlit
openai-whisper
certifi
